feat: UIStateを構築するロジックを切り出しました
