+++ modules/features/note/src/main/java/net/pantasystem/milktea/note/reaction/history/ReactionHistoryListAdapter.kt
@@ -13,10 +13,10 @@ import net.pantasystem.milktea.common.glide.GlideApp
 import net.pantasystem.milktea.model.account.Account
 import net.pantasystem.milktea.model.notes.reaction.LegacyReaction
 import net.pantasystem.milktea.model.user.User
+import net.pantasystem.milktea.note.EmojiType
 import net.pantasystem.milktea.note.R
 import net.pantasystem.milktea.note.databinding.ItemReactionHistoryHeaderBinding
 import net.pantasystem.milktea.note.databinding.ItemSimpleUserBinding
-import net.pantasystem.milktea.note.reaction.viewmodel.EmojiType
 
 class ReactionHistoryListAdapter(
     val lifecycleOwner: LifecycleOwner
